Move "supsub" handlers to functions/subsub.js.
