+++ src/functions.js
@@ -36,6 +36,7 @@ import "./functions/sizing";
 import "./functions/smash";
 import "./functions/sqrt";
 import "./functions/styling";
+import "./functions/supsub";
 import "./functions/symbolsOp";
 import "./functions/symbolsOrd";
 import "./functions/symbolsSpacing";
